Positivos

+++ Tema2/PositivosPrueba.java
@@ -1,6 +1,6 @@
 public class PositivosPrueba {
     public static void main(String[] args) {
-        double resultado = Positivos.positivos();
+        int resultado = Positivos.positivos();
         System.out.println(resultado);
     }
 }
